Fix undefined index minify (#554)
* fix for PHP Notice:  Undefined index: minify in CacheFlush_Locally.php on line 84

* Update readme

+++ CacheFlush_Locally.php
@@ -81,7 +81,7 @@ class CacheFlush_Locally {
 	}
 
 	function minifycache_flush_all( $extras = array() ) {
-		if ( $extras['minify'] == 'purge_map' )
+		if ( isset( $extras['minify'] ) && $extras['minify'] == 'purge_map' )
 			delete_option( 'w3tc_minify' );
 
 		$this->minifycache_flush( $extras );
